requirements.txt
# Dependency for UI 
streamlit
watchdog

# Backend dependencies
python-dotenv
fastapi[standard]
uvicorn

# Langchain dependencies
langchain_openai
langchain
langgraph

# Add these updated dependencies to requirements.txt
# Vector store dependencies
pinecone-client>=3.0.0
langchain-pinecone
langchain-community
datasets  
langchain-pinecone      
langchain-community     
